+++ api/controllers/config/breaks/set-remote.js
@@ -43,7 +43,7 @@ module.exports = {
       // Modify config
       for (var key in inputs) {
         if (Object.prototype.hasOwnProperty.call(inputs, key)) {
-          sails.config.custom.specialBreaks.remote[key] = inputs[key]
+          sails.config.custom.specialBreaks.remote[key] = await sails.helpers.break.removeNullTasks(inputs[key])
         }
       }